+++ CHANGELOG.md
@@ -12,6 +12,7 @@ That's all, matcher will stop when detects next line started with `###` match
 ### 4.6.0
 * Now players who's in a game can't swap items to off hand
 * Players now can't drag special items in theirs inventory
+* Leaves will no longer be decayed if this is within an arena.
 
 ### 4.5.9 (28.03.2021 - 13.04.2021)
 * Fix when Splash_potion material not existed on 1.8 servers